feat(src): 修改线程池注册器的相关逻辑以及进行公共行为抽象
